# noncliff:  improve performance of  `_allthreesumtozero`