@Fe-r-oz 2024-10-04
I think the PR is ready for review. Thanks for your tip about using `evals=1 `and `setup` from some time ago, that helped me from not producing misleading results... Thank you!